feat: Update configuration poll interval default value
The changes update the default value for the `CONFIGURATION_POLL_INTERVAL`
environment variable. The previous implementation used `Number.parseInt`
which could result in a value of `0` if the environment variable was not
set. The new implementation uses `Number()` which will return `NaN` if the
environment variable is not set, and then falls back to a default value of
60000 (1 minute).